Loosen extension check for CRD verify script
The script to verify CRD codegen checks diffs only against files that
end in exactly `.crd.yaml`. This loosens that to just `crd.yaml` (with
no leading `.`) to ensure all CRDs are captured.

There are very few English words ending in `crd`, so I don't think there
is a risk of this conflicting.