Prefer OS_GIT_VERSION over SOURCE_GIT_TAG when getting build version
When ART builds an image, it copies the content of the source (github) repo into a production repo, and builds from that. The added SOURCE_GIT_* vars in Dockerfile are intended to relay the state of the *original* source repo, in case that is important at build or runtime.

To get the production version, one needs to read OS_GIT_VERSION instead.