Try to only strip local path prefixes _within_ an argument
Given a log line like
> /abc/github.com/openshift/build-machinery-go/abc/ /github.com/openshift/build-machinery-go/ /def/github.com/openshift/build-machinery-go/def/
currently the edit command produces
> /github.com/openshift/build-machinery-go/def/
matching everything from the first / to the end of the line,
obscuring the command and hiding possibly relevant changes.

Use [^ ]* instead of .* to at least _try to_ treat one path at a time;
it might not be perfect but it's an improvement.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>