ensure imagebuilder
make/targets/openshift/images.mk references imagebuilder, which I didn't
have. There are other convenient `ensure-*` targets, so I thought I
would make one for this, and add it as a requisite for the generated
`image-*` rules.

However, `imagebuilder` is a go tool, and as such should be installed
via `go get`, which behaves unreliably/inconsistently depending on the
go version, variables, and dependencies in the consuming environment.
Therefore, rather than trying to install it automatically, we instead
produce a useful error message.

If we find the tool is installed, we check the version and emit a
warning if it doesn't match what's expected.