commitchecker: valiate merge-base, commits after that
Every time we use commitchecker today in a rebase PR, we need to
manually override it as it's validating every commit message, even the
ones that come from upstream. This PR adds a new capability to the
commitchecker, allowing it to load configuration from the repository
it's validating in order to determine the merge-base, validate it, and
only validate UPSTREAM commits after that point.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>