Deprecate controller-gen.mk
Though well-intentioned, controller-gen.mk is tough to maintain due to
the need to build and publish binaries for multiple platforms at any
desired release.

This PR does not change functionality, but leaves a deprecation comment
in the source file with the idea that current consumers wishing to
upgrade their controller-gen will find it after running into the same
problems I did (see the issue).

Resolves #93